feat(batteries): Rename `node` to `batteries`
`node` is confusing in the context of a DOM. The new name makes it easier to figure out what is happening.
